+++ test/test_adau1361.cpp
@@ -3,23 +3,26 @@
 
 #include "codec/adau1361.hpp"
 #include "codec/umbadau1361lower.hpp"
-#include "i2c/i2cmasterinterface.hpp"
+#include "i2c/i2cmaster.hpp"
 
 class Adau1361Test : public ::testing::Test {
  protected:
   virtual void SetUp() {
-    codec_lower_ = new ::rpp_driver::MockAdau1361Lower(i2c_);
+    i2c_ = new ::rpp_driver::MockI2cMaster(sdk_);
+    codec_lower_ = new ::rpp_driver::MockAdau1361Lower(*i2c_);
     codec_ = new ::rpp_driver::Adau1361(fs_, master_clk_, *codec_lower_);
   }
 
   virtual void TearDown() {
     delete codec_;
     delete codec_lower_;
+    delete i2c_;
   }
 
   uint fs_ = 48000;
   uint master_clk_ = 12000000;
-  ::rpp_driver::MockI2cMasterInterface i2c_;
+  ::rpp_driver::SdkWrapper sdk_;
+  ::rpp_driver::MockI2cMaster *i2c_;
   ::rpp_driver::MockAdau1361Lower *codec_lower_;
   ::rpp_driver::Adau1361 *codec_;
 };