+++ test/test_mockadau1361.cpp
@@ -13,16 +13,19 @@ using ::testing::_;
 class MockAdau1361Test : public ::testing::Test {
  protected:
   virtual void SetUp() {
-    codec_lower_ = new ::rpp_driver::MockAdau1361Lower(i2c_);
+    i2c_ = new ::rpp_driver::MockI2cMaster(sdk_);
+    codec_lower_ = new ::rpp_driver::MockAdau1361Lower(*i2c_);
     codec_ = new ::rpp_driver::MockAdau1361(*codec_lower_);
   }
 
   virtual void TearDown() {
     delete codec_;
     delete codec_lower_;
+    delete i2c_;
   }
 
-  ::rpp_driver::MockI2cMasterInterface i2c_;
+  ::rpp_driver::SdkWrapper sdk_;
+  ::rpp_driver::MockI2cMaster *i2c_;
   ::rpp_driver::MockAdau1361Lower *codec_lower_;
   ::rpp_driver::Adau1361 *codec_;
 };  // MockAdau1361Test